add:FlyerStateWanderParam ctor && fix:requested changes
